+++ src/main/java/cart/controller/api/ProductController.java
@@ -1,8 +1,8 @@
-package cart.controller;
+package cart.controller.api;
 
 import cart.dto.request.CreateProductRequest;
 import cart.dto.request.UpdateProductRequest;
-import cart.service.CartService;
+import cart.service.ProductService;
 import java.net.URI;
 import javax.validation.Valid;
 import org.springframework.beans.factory.annotation.Autowired;
@@ -17,28 +17,28 @@ import org.springframework.web.bind.annotation.RestController;
 @RestController
 public class ProductController {
 
-    private final CartService cartService;
+    private final ProductService productService;
 
     @Autowired
-    public ProductController(final CartService cartService) {
-        this.cartService = cartService;
+    public ProductController(final ProductService productService) {
+        this.productService = productService;
     }
 
     @PostMapping("/products")
     public ResponseEntity<Void> createProduct(@RequestBody @Valid final CreateProductRequest createProductRequest) {
-        cartService.insert(createProductRequest);
+        productService.insert(createProductRequest);
         return ResponseEntity.created(URI.create("/products")).build();
     }
 
     @PutMapping("/products/{id}")
     public ResponseEntity<Void> updateProduct(@PathVariable("id") final Long id, @RequestBody @Valid final UpdateProductRequest updateProductRequest) {
-        cartService.update(id, updateProductRequest);
+        productService.update(id, updateProductRequest);
         return ResponseEntity.ok().build();
     }
 
     @DeleteMapping("/products/{id}")
     public ResponseEntity<Void> deleteProduct(@PathVariable("id") final Long id) {
-        cartService.delete(id);
+        productService.delete(id);
         return ResponseEntity.accepted().build();
     }
 }
